add service and error handling
