+++ inventory/utils.py
@@ -1,16 +1,3 @@
-from inventory.models import Material, MaterialStock, MaterialQuantity, Store
-
-
-def get_restock_total_price(materials):
-    total_price = 0
-    for material in materials:
-        if "material" in material or "quantity" in material:
-            price = Material.objects.get(pk=material.get('material')).price
-            total_price += price * material.get('quantity')
-
-    return round(total_price, 2)
-
-
 def get_model_obj_property(model, instance, exclude=("pk")):
     field_names = [f.name for f in model._meta.fields]
     property_names = [
@@ -18,28 +5,3 @@ def get_model_obj_property(model, instance, exclude=("pk")):
         if isinstance(getattr(model, name), property) and name not in exclude
     ]
     return dict((name, getattr(instance, name)) for name in field_names + property_names)
-
-
-def get_product_remaining_capacities(obj):
-    data = []
-    if not isinstance(obj, Store):
-        return data
-    for product in obj.products.all():
-        material_quantities_need = MaterialQuantity.objects.filter(
-            product=product
-        )
-        material_quantity_list = []
-        for material_quantity in material_quantities_need:
-            quantity_needed = material_quantity.quantity
-            material_stock = obj.material_stocks.get(
-                material=material_quantity.ingredient
-            )
-            current_quantity = material_stock.current_capacity
-            material_quantity_list.append(
-                int(current_quantity / quantity_needed)
-            )
-        data.append({
-            'product': product.product_id,
-            'quantity': min(material_quantity_list, default=0),
-        })
-    return data
